% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_days_out.R
\name{list_days_out}
\alias{list_days_out}
\title{Get a list of outage details by plant}
\usage{
list_days_out(
  start_date = "2024-01-01 00:00:00",
  end_date = "2024-12-31 23:59:59",
  reason = NA,
  percent_min = 0,
  mw_min = 0,
  num = 50
)
}
\arguments{
\item{start_date}{A date in string form "YYYY-MM-DD"}

\item{end_date}{A date in string form "YYYY-MM-DD"}

\item{reason}{A specific outage reason given.}

\item{percent_min}{A threshold; don't show outages below this threshold.}

\item{mw_min}{A threshold; don't show outages below this threshold.}

\item{num}{Number of outages to show.}
}
\description{
Get a list of outage details by plant
}
\examples{
# To see possible outage reasons:
unique(outage$`NATURE OF WORK`)


}
