% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_from_source.R
\name{load_data}
\alias{load_data}
\title{Loads all necessary datasets
Object names:  DAM, RTM, interchange, outage, weather}
\usage{
load_data(weather = FALSE)
}
\arguments{
\item{weather}{Should we load the large (2GB) weather dataset?}
}
\description{
Loads all necessary datasets
Object names:  DAM, RTM, interchange, outage, weather
}
