% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_from_source.R
\name{load_weather_data}
\alias{load_weather_data}
\title{Loads all necessary datasets
Overwrites global object \code{weather}}
\usage{
load_weather_data()
}
\description{
Loads all necessary datasets
Overwrites global object \code{weather}
}
