% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outages.R
\name{plot_all_outages}
\alias{plot_all_outages}
\title{Plot all plant outages in date range, by MWH and percent}
\usage{
plot_all_outages(
  start_date = "2024-01-01",
  end_date = "2025-01-01",
  percent_cutoff = 50,
  interactive = TRUE
)
}
\arguments{
\item{start_date}{A date in string form "YYYY-MM-DD"}

\item{end_date}{A date in string form "YYYY-MM-DD"}

\item{percent_cutoff}{Filter to outages above a threshold.}

\item{interactive}{Static plot or interactive html plot?}
}
\description{
Plot all plant outages in date range, by MWH and percent
}
\examples{
# to see available plant names:
outage |> pull(`RESOURCE NAME`) |> unique()


}
