% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exports.R
\name{plot_exports}
\alias{plot_exports}
\title{Plot export trends over date range}
\usage{
plot_exports(
  start_date = "2024-01-01",
  end_date = "2025-01-01",
  interactive = TRUE
)
}
\arguments{
\item{start_date}{A date in string form "YYYY-MM-DD"}

\item{end_date}{A date in string form "YYYY-MM-DD"}

\item{interactive}{Static plot or interactive html plot?}
}
\description{
Plot export trends over date range
}
