% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outages.R
\name{plot_forced_curtailments}
\alias{plot_forced_curtailments}
\title{Plot all forced curtailments in date range, by MWH and percent}
\usage{
plot_forced_curtailments(
  start_date = "2024-01-01",
  end_date = "2025-01-01",
  interactive = TRUE
)
}
\arguments{
\item{start_date}{A date in string form "YYYY-MM-DD"}

\item{end_date}{A date in string form "YYYY-MM-DD"}

\item{interactive}{Static plot or interactive html plot?}
}
\description{
Plot all forced curtailments in date range, by MWH and percent
}
