% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outages.R
\name{plot_plant_outage}
\alias{plot_plant_outage}
\title{Plot plant outages in date range for specific plant}
\usage{
plot_plant_outage(
  plant_name,
  start_date = "2024-01-01",
  end_date = "2025-01-01",
  interactive = TRUE
)
}
\arguments{
\item{plant_name}{The name of a plant in the dataset.}

\item{start_date}{A date in string form "YYYY-MM-DD"}

\item{end_date}{A date in string form "YYYY-MM-DD"}

\item{interactive}{Static plot or interactive html plot?}
}
\description{
Plot plant outages in date range for specific plant
}
\examples{
# to see available plant names:
unique(outage$`RESOURCE NAME`)


}
